Correct likelihood ratio calculation function and document
1. There was a typo in the math for the expected value of cell a,
introduced in 1cccafa.

2. This calculates the "short" version of the log-likelihood, summing
only over the cells containing the count of the target word. There's no
reason to think that is chi^2-distributed, though it appears to be
commonly used as a summary statistic. Default to the full log-likelihood
so the statistic can be used in tests, such as in keyness_pairs() and
keyness_table().